Fixed compile warnings specific for 64-bit architectures
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@3343 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/mi_xmlrpc/abyss_response.c
@@ -225,7 +225,7 @@ ResponseContentLength(TSession * const sessionP,
                       uint64_t   const len) {
     char contentLengthValue[32];
     
-    sprintf(contentLengthValue, "%llu", len);
+    sprintf(contentLengthValue, "%llu", (long long unsigned int)len);
 
     return ResponseAddField(sessionP, "Content-length", contentLengthValue);
 }
